Split reporter in multiple reporters.
This change makes this library more agnostic
to external metrics providers.

It allows users to plug their own reporters if they don't
use statsd.
