+++ lib/trashed/railtie.rb
@@ -4,7 +4,7 @@ require 'trashed/reporter'
 
 module Trashed
   class Railtie < ::Rails::Railtie
-    config.trashed = Trashed::Reporter.new
+    config.trashed = Trashed::Reporter::Aggregator.new
 
     # Middleware would like to emit tagged logs after Rails::Rack::Logger
     # pops its tags. Introduce this haxware to stash the tags in the Rack
@@ -22,11 +22,7 @@ module Trashed
     end
 
     initializer 'trashed' do |app|
-      require 'statsd'
-
-      app.config.trashed.timing_sample_rate ||= 0.1
-      app.config.trashed.gauge_sample_rate ||= 0.05
-      app.config.trashed.logger ||= Rails.logger
+      app.config.trashed.add_reporter Trashed::Reporter::Logger.new(Rails.logger)
 
       app.middleware.insert_after ::Rack::Runtime, Trashed::Rack, app.config.trashed
       app.middleware.insert_after ::Rails::Rack::Logger, ExposeLoggerTagsToRackEnv
