+++ test/reporter_test.rb
@@ -4,15 +4,9 @@ require 'logger'
 require 'stringio'
 
 class ReporterTest < Minitest::Test
-  def setup
-    @reporter = Trashed::Reporter.new
-    @reporter.timing_sample_rate = 1
-    @reporter.gauge_sample_rate = 1
-  end
-
   def test_sample_rate_defaults
-    assert_equal 0.1, Trashed::Reporter.new.timing_sample_rate
-    assert_equal 0.05, Trashed::Reporter.new.gauge_sample_rate
+    assert_equal 0.1, Trashed::Reporter::Statsd.new.timing_sample_rate
+    assert_equal 0.05, Trashed::Reporter::Statsd.new.gauge_sample_rate
   end
 
   def test_report_logger
@@ -42,21 +36,24 @@ class ReporterTest < Minitest::Test
   end
 
   def test_tagged_logger
-    @reporter.logger = logger = Logger.new(out = StringIO.new)
+    out = StringIO.new
+    logger = Logger.new(out)
     class << logger
       attr_reader :tags
       def tagged(tags) @tags = tags; yield end
     end
 
-    @reporter.report_logger 'trashed.logger.tags' => %w(a b c), 'trashed.timings' => { :'Time.wall' => 1 }
+    reporter = Trashed::Reporter::Logger.new(logger)
+
+    reporter.report 'trashed.logger.tags' => %w(a b c), 'trashed.timings' => { :'Time.wall' => 1 }
     assert_match 'Rack handled in 1.00ms.', out.string
     assert_equal %w(a b c), logger.tags
   end
 
   private
   def assert_report_logs(string, timings = {})
-    @reporter.logger = Logger.new(out = StringIO.new)
-    @reporter.report_logger 'trashed.timings' => timings.merge(:'Time.wall' => 1)
+    logger = Trashed::Reporter::Logger.new(Logger.new(out = StringIO.new))
+    logger.report 'trashed.timings' => timings.merge(:'Time.wall' => 1)
     assert_match string, out.string
   end
 end
